Rename translation.segments.html module to translations.strings
