+++ wagtail_localize/translation/segments/ingest.py
@@ -6,7 +6,7 @@ from wagtail.core import blocks
 from wagtail.core.fields import RichTextField, StreamField
 from wagtail.core.rich_text import RichText
 
-from .html import restore_strings
+from ..strings import restore_strings
 
 
 def organise_template_segments(segments):
