+++ wagtail_localize/translation/segments/tests/test_segment_ingestion.py
@@ -13,8 +13,8 @@ from wagtail_localize.translation.segments import (
     TemplateValue,
     RelatedObjectValue,
 )
-from wagtail_localize.translation.segments.html import String
 from wagtail_localize.translation.segments.ingest import ingest_segments
+from wagtail_localize.translation.strings import String
 
 
 def make_test_page(**kwargs):
