+++ wagtail_localize/translation/tests/test_translationsource.py
@@ -22,7 +22,7 @@ from wagtail_localize.translation.models import (
 )
 from wagtail_localize.translation.segments import TemplateValue, RelatedObjectValue
 from wagtail_localize.translation.segments.extract import extract_segments
-from wagtail_localize.translation.segments.html import String
+from wagtail_localize.translation.strings import String
 
 
 def insert_segments(revision, locale, segments):
